Who you gonna call? Ghostbusters!

+++ Userland/SampleCodeModule/c/testing_utils.c
@@ -69,7 +69,7 @@ void endless_loop(){
 
 void endless_loop_print()
 {
-  uint8_t pid = sys_get_pid();
+  uint32_t pid = sys_get_pid();
 
   while(1){
     printf("%d ", pid);
